flask-project/pyproject.toml
[tool.poetry]
name = "flask-project"
version = "0.1.0"
description = ""
authors = ["rede.akbar <ribonred@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flasgger = "^0.9.7.1"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
sqlalchemy-serializer = "^1.4.22"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
